Fix parameters and adjust on push skill

+++ const/constants.cpp
@@ -59,7 +59,7 @@ MRCConstants::MRCConstants(QString configFileName, bool playingAgainstWarthog)
     std::cout<< MRCConstants::defaultBold << "NumberOfPlayers: "<< MRCConstants::green <<_qtPlayers<<std::endl<< MRCConstants::reset;
 
     _maxKickPower = doc_map["MaxKickPower"].toFloat();
-    std::cout<< MRCConstants::defaultBold << "MaxKickPower: "<< MRCConstants::green <<doc_map["MaxKickPower"].toInt()<<std::endl<< MRCConstants::reset;
+    std::cout<< MRCConstants::defaultBold << "MaxKickPower: "<< MRCConstants::green <<doc_map["MaxKickPower"].toFloat()<<std::endl<< MRCConstants::reset;
 
     _robotRadius = doc_map["RobotRadius"].toFloat();
     std::cout<< MRCConstants::defaultBold << "RobotRadius: "<< MRCConstants::green <<_robotRadius<<std::endl<< MRCConstants::reset;
